Simplify api / repo
